Fix noise test

+++ test/test_noise.dart
@@ -10,14 +10,14 @@ class NoiseTest extends BaseTest {
 
     // Cache several values at known coordinates
     for (var i = 0.0; i < values2D.length; ++i) {
-      values2D[i] = noise.noise2D(i, i);
-      values3D[i] = noise.noise3D(i, i, i);
+      values2D[i.toInt()] = noise.noise2D(i, i);
+      values3D[i.toInt()] = noise.noise3D(i, i, i);
     }
 
     // Ensure that querying those same coordinates repeats the cached value
     for (var i = 0.0; i < values2D.length; ++i) {
-      expect(values2D[i], equals(noise.noise2D(i, i)));
-      expect(values3D[i], equals(noise.noise3D(i, i, i)));
+      expect(values2D[i.toInt()], equals(noise.noise2D(i, i)));
+      expect(values3D[i.toInt()], equals(noise.noise3D(i, i, i)));
     }
   }
